# api/swagger/v1/simulation-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Simulation service",
    "version": "1.0",
    "contact": {
      "name": "simulation project",
      "url": "https://github.com/olamai/simulation",
      "email": "zacharyholland@gmail.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/agent": {
      "post": {
        "summary": "Create new agent",
        "operationId": "CreateAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAgentResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAgentRequest"
            }
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/agent/{id}": {
      "get": {
        "summary": "Delete an agent",
        "operationId": "DeleteAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteAgentResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Unique integer identifier of the agent to delete",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SimulationService"
        ]
      },
      "put": {
        "summary": "Perform an action for an agent",
        "operationId": "ExecuteAgentAction",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ExecuteAgentActionResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Id for the agent that should perform the action",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ExecuteAgentActionRequest"
            }
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/agent/{id}/observation": {
      "get": {
        "summary": "Get the observation for an agent",
        "operationId": "GetAgentObservation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAgentObservationResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Agent that should perform the action",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/entity/{id}": {
      "get": {
        "summary": "Get data for an entity",
        "operationId": "GetEntity",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEntityResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Unique integer identifier of the agent",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/reset": {
      "put": {
        "summary": "Reset the world",
        "operationId": "ResetWorld",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ResetWorldResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ResetWorldRequest"
            }
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/spectator": {
      "get": {
        "summary": "Create a new spectator stream",
        "operationId": "CreateSpectator",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/v1CellUpdate"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "description": "API versioning: it is my best practice to specify version explicitly.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "description": "Id of the spectator.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    },
    "/v1/spectator/{id}/subscribeToRegion": {
      "put": {
        "summary": "Subscribe the spectator of given id to a region",
        "operationId": "SubscribeSpectatorToRegion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SubscribeSpectatorToRegionResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Id of the spectator",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SubscribeSpectatorToRegionRequest"
            }
          }
        ],
        "tags": [
          "SimulationService"
        ]
      }
    }
  },
  "definitions": {
...
    "v1Action": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "identifier for the action\nEx. Move, Eat, etc."
        },
        "direction": {
          "type": "string",
          "description": "direction to perform the action in\nEx. Up, Down, etc."
        }
      }
    },
...
    "v1CreateAgentRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "agent": {
          "$ref": "#/definitions/v1Entity",
          "title": "Agent to add"
        }
      },
      "title": "----------------------\n-- Agent\n----------------------\nRequest data to create new agent"
    },
    "v1CreateAgentResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "id": {
          "type": "string",
          "format": "int64",
          "title": "ID of created agent"
        }
      },
      "title": "Contains data of created agent"
    },
    "v1DeleteAgentResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "deleted": {
          "type": "string",
          "format": "int64",
          "title": "Contains number of entities have beed deleted\nEquals 1 in case of successful delete"
        }
      },
      "title": "Contains status of delete operation"
    },
    "v1Entity": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Unique integer identifier of the agent"
        },
        "x": {
          "type": "integer",
          "format": "int32",
          "title": "Position"
        },
        "y": {
          "type": "integer",
          "format": "int32"
        },
        "class": {
          "type": "string",
          "title": "Entity class"
        }
      },
      "title": "Taks we have to do"
    },
    "v1ExecuteAgentActionRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Id for the agent that should perform the action"
        },
        "action": {
          "$ref": "#/definitions/v1Action",
          "title": "Action id"
        }
      }
    },
    "v1ExecuteAgentActionResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "isAgentStillAlive": {
          "type": "boolean",
          "format": "boolean"
        },
        "wasActionSuccessful": {
          "type": "boolean",
          "format": "boolean",
          "title": "Was the agent able to perform the action"
        }
      }
    },
    "v1GetAgentObservationResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "observation": {
          "$ref": "#/definitions/v1Observation",
          "title": "Was the agent able to perform the action"
        }
      }
    },
    "v1GetEntityResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "entity": {
          "$ref": "#/definitions/v1Entity",
          "title": "Task entity read by ID"
        }
      },
      "title": "Contains entity data specified in by ID request"
    },
    "v1Observation": {
      "type": "object",
      "properties": {
        "Cells": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Cells around the agent"
        },
        "Alive": {
          "type": "boolean",
          "format": "boolean",
          "title": "Agent stats"
        },
        "Energy": {
          "type": "integer",
          "format": "int32"
        },
        "Health": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1Region": {
      "type": "object",
      "properties": {
        "x": {
          "type": "integer",
          "format": "int32"
        },
        "y": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ResetWorldRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string"
        }
      }
    },
    "v1ResetWorldResponse": {
      "type": "object"
    },
    "v1SubscribeSpectatorToRegionRequest": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "id": {
          "type": "string",
          "title": "Id of the spectator"
        },
        "region": {
          "$ref": "#/definitions/v1Region",
          "title": "What region to add to"
        }
      }
    },
    "v1SubscribeSpectatorToRegionResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API versioning: it is my best practice to specify version explicitly"
        },
        "successful": {
          "type": "boolean",
          "format": "boolean",
          "title": "Success status"
        }
      }
    }
  },
...
}